queue class works by itself in its own file
