+++ data-structures/stacksAndQueues/__tests__/queue.test.js
@@ -1,0 +1,3 @@
+'use strict';
+
+require('../model/queue.js');
