+++ data-structures/stacksAndQueues/__tests__/stack.test.js
@@ -1,0 +1,3 @@
+'use strict';
+
+require('../model/stack.js');
