+++ data-structures/stacksAndQueues/model/node.js
@@ -1,0 +1,10 @@
+'use strict';
+
+class Node {
+  constructor(value) {
+    this.value = value;
+    this.next = null;
+  }
+}
+
+module.exports = Node;
